test: tests have been added for Postman Collection
