+++ tests/TestingTest.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests;
+
+use Lion\Test\Test;
+
+class TestingTest extends Test
+{
+    public function test(): void
+    {
+        $this->expectNotToPerformAssertions();
+    }
+}
